requirements.txt
hydra-core==1.0.0
imbalanced-learn==0.7.0
mlflow==1.11.0
numpy==1.16.4
omegaconf==2.0.1
pandas==1.1.2
parso==0.8.0
scikit-learn==0.23.2
scipy==1.1.0
skorch==0.9.0
torch==1.1.0
